better project structure, separating server
