+++ statki-game-server/config/gameSettings.js
@@ -1,0 +1,4 @@
+module.exports = Object.freeze({
+  WIDTH: 0,
+  HEIGHT: 0,
+});
