+++ statki-game-server/package.json
@@ -2,7 +2,7 @@
   "name": "game-server",
   "version": "1.0.0",
   "description": "TeamShips game-server.",
-  "main": "server.js",
+  "main": "game.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
