+++ statki/src/main/resources/application.yml
@@ -3,6 +3,6 @@ app:
     jdbc-url: jdbc:postgresql://localhost:5432/statkidb
     username: postgres
 #    password: 1234
-#    password: Statki123
-    password: Poirot99
+    password: Statki123
+#    password: Poirot99
     pool-size: 1000
